Makes loading of data cross-os compatible
Replaces the function to read data, with the pathlib module to
avoid hardcoding the directory of the images. Now looks into the
data directory within the git repo - which given the size of the data
is now included in the gitignore, so one will have to manually download
and place these data into the directory.